backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-pymongo==2.3.0
pyjwt==2.1.0
werkzeug==2.0.1
face-recognition==1.3.0
python-dotenv==0.19.0
numpy==1.24.3
opencv-python>=4.7.0.72
flask-mail==0.9.1